requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
# faiss -cpu
sentence-transformers
python-dotenv
pypdf